Only use cecil reader
