+++ src/OpenRiaServices.Tools/Framework/SourceLocation/PdbSourceFileProviderFactory.cs
@@ -1,39 +1,43 @@
-﻿namespace OpenRiaServices.Tools.SourceLocation
-{
-    using System;
-    using System.Collections.Generic;
-    using System.Diagnostics;
-    using System.Diagnostics.SymbolStore;
-    using System.Globalization;
-    using System.Reflection;
-    using OpenRiaServices;
-    using OpenRiaServices.Tools.Pdb.SymStore;
-    using OpenRiaServices.Tools.SharedTypes;
+﻿using System;
+using System.Collections.Generic;
+using System.Diagnostics;
+using System.Diagnostics.SymbolStore;
+using System.Globalization;
+using System.IO;
+using System.Linq;
+using System.Reflection;
+using Mono.Cecil;
+using Mono.Cecil.Cil;
+using Mono.Cecil.Pdb;
+using OpenRiaServices.Tools.SharedTypes;
 
+namespace OpenRiaServices.Tools.SourceLocation
+{
     /// <summary>
     /// PDB-based implementation of <see cref="ISourceFileProvider"/> to locate source files for types or methods.
     /// </summary>
     internal class PdbSourceFileProviderFactory : ISourceFileProviderFactory
     {
         private readonly ILogger _logger;
-        private readonly string _symbolSearchPath;
+        private readonly string[] _symbolSearchPaths;
 
         /// <summary>
         /// Creates a new instance.
         /// </summary>
-        /// <param name="symbolSearchPath">Optional list of semicolon separated paths to search for PDB files.</param>
+        /// <param name="symbolSearchPaths">Optional list of paths to search for PDB files.</param>
         /// <param name="logger">Optional logger to use to report errors or warnings.</param>
-        public PdbSourceFileProviderFactory(string symbolSearchPath, ILogger logger)
+        public PdbSourceFileProviderFactory(string[] symbolSearchPaths, ILogger logger)
             : base()
         {
             this._logger = logger;
-            this._symbolSearchPath = symbolSearchPath;
+            this._symbolSearchPaths = symbolSearchPaths;
         }
+
 
         #region ISourceFileProviderFactory
         public ISourceFileProvider CreateProvider()
         {
-            return new PdbSourceFileProvider(this._symbolSearchPath, this._logger);
+            return new CecilSourceFileProvider(this._symbolSearchPaths, this._logger);
         }
         #endregion // ISourceFileProviderFactory
 
@@ -43,122 +47,90 @@
         /// of a type.  This class is instantiated when the base class asks
         /// for a location context, and then it is disposed by the base.
         /// </summary>
-        internal class PdbSourceFileProvider : ISourceFileProvider, IDisposable
+        internal sealed class CecilSourceFileProvider : ISourceFileProvider
         {
             private readonly ILogger _logger;
-            private readonly string _symbolSearchPath;
-            private readonly Dictionary<Assembly, ISymbolReader> _symbolReadersByType = new Dictionary<Assembly, ISymbolReader>();
+            private readonly Dictionary<string, AssemblyDefinition> _assembliesByLocation = new();
+            private readonly ISymbolReaderProvider _symbolReaderProvider;
 
-            internal PdbSourceFileProvider(string symbolSearchPath, ILogger logger)
+            internal CecilSourceFileProvider(string[] symbolSearchPath, ILogger logger)
             {
-                this._symbolSearchPath = symbolSearchPath;
+                _symbolReaderProvider = new SymbolReaderProvider(symbolSearchPath);
                 this._logger = logger;
             }
 
             /// <summary>
-            /// Indexer that gets the <see cref="ISymbolReader"/> appropriate to analyze
-            /// the specified <paramref name="type"/>.  This indexer caches by Assembly.
+            /// Indexer that gets the <see cref="TypeDefinition"/> appropriate to analyze
+            /// the specified <paramref name="type"/>.
             /// </summary>
             /// <param name="type">The type we will analyze.</param>
-            /// <returns>A <see cref="ISymbolReader"/> or <c>null</c> if one is not available.</returns>
-            internal ISymbolReader this[Type type]
+            /// <returns>A <see cref="TypeDefinition"/> or <c>null</c> if one is not available.</returns>
+            internal TypeDefinition this[Type type]
             {
                 get
                 {
                     Debug.Assert(type != null, "The type is required");
                     Assembly assembly = type.Assembly;
-                    ISymbolReader reader = null;
+                    string location = assembly.Location;
 
                     // Lazily create the readers for assemblies we have not yet encountered.
-                    if (!this._symbolReadersByType.TryGetValue(assembly, out reader))
+                    if (!this._assembliesByLocation.TryGetValue(location, out AssemblyDefinition assemblyDefinition))
                     {
                         // We don't create symbol readers for System assemblies
                         if (!assembly.IsSystemAssembly())
                         {
                             // Lazy create.  Note that a null is a legal result and will
                             // be cached to avoid redundant failures
-                            reader = this.CreateSymbolReader(assembly);
+                            try
+                            {
+                                var parameters = new ReaderParameters
+                                {
+                                    ReadSymbols = true,
+                                    ReadWrite = false,
+                                    SymbolReaderProvider = _symbolReaderProvider
+                                };
+
+                                assemblyDefinition = AssemblyDefinition.ReadAssembly(location, parameters);
+                            }
+                            catch (Exception ex)
+                            {
+                                _logger.LogWarning($"Failed to load metadata for {assembly.FullName}: {ex}");
+                                // Cache null
+                                assemblyDefinition = null;
+                            }
                         }
 
-                        this._symbolReadersByType[assembly] = reader;
+                        this._assembliesByLocation[location] = assemblyDefinition;
                     }
 
-                    return reader;
+                    return assemblyDefinition?.MainModule.GetType(type.Namespace, type.Name);
                 }
             }
 
             /// <summary>
             /// Returns the name of the file for the given method using the given symbol reader
             /// </summary>
-            /// <param name="reader">The reader to use</param>
+            /// <param name="typeDefinition">The type tos search in</param>
             /// <param name="methodBase">The method to lookup</param>
             /// <returns>The file containing the method or null.</returns>
-            private static string GetFileForMethod(ISymbolReader reader, MethodBase methodBase)
+            private static string GetFileForMethod(TypeDefinition typeDefinition, MethodBase methodBase)
             {
+                // Token is unique identifer per compilation (so cannot use old pdb)
                 int token = methodBase.MetadataToken;
 
-                ISymbolMethod methodSymbol = reader == null ? null : reader.GetMethod(new SymbolToken(token));
-                if (methodSymbol != null)
-                {
-                    int count = methodSymbol.SequencePointCount;
-                    if (count == 0)
-                        return null;
-
-                    // Get the sequence points from the symbol store. 
-                    // We could cache these arrays and reuse them.
-                    int[] offsets = new int[count];
-                    ISymbolDocument[] docs = new ISymbolDocument[count];
-                    int[] startColumn = new int[count];
-                    int[] endColumn = new int[count];
-                    int[] startRow = new int[count];
-                    int[] endRow = new int[count];
-                    methodSymbol.GetSequencePoints(offsets, docs, startRow, startColumn, endRow, endColumn);
-
-                    return docs[0].URL.ToString();
-                }
-                return null;
+                var method = typeDefinition.Methods.FirstOrDefault(md => md.MetadataToken.ToInt32() == token);
+                return method?.DebugInformation.SequencePoints.FirstOrDefault()?.Document.Url;
             }
 
-            /// <summary>
-            /// Creates a <see cref="ISymbolReader"/> for the given <paramref name="assembly"/>.
-            /// </summary>
-            /// <param name="assembly">The assembly whose reader is needed.</param>
-            /// <returns>The <see cref="ISymbolReader"/> instance or <c>null</c> if one cannot be created (e.g. no PDB).</returns>
-            private ISymbolReader CreateSymbolReader(Assembly assembly)
-            {
-                Debug.Assert(assembly != null, "The assembly is required");
-                ISymbolReader reader = null;
-                string assemblyFile = assembly.Location;
-
-                try
-                {
-                    reader = SymbolAccess.GetReaderForFile(assemblyFile, this._symbolSearchPath);
-                }
-                catch (System.Runtime.InteropServices.COMException cex)
-                {
-                    // Experience has shown some large PDB's can exhaust memory and cause COM failures.
-                    // When this occurs, we log a warning and continue as if the PDB was not there.
-                    if (this._logger != null)
-                    {
-                        this._logger.LogWarning(string.Format(CultureInfo.CurrentCulture, Resource.Failed_To_Open_PDB, assemblyFile, cex.Message));
-                    }
-                }
-
-                return reader;
-            }
 
             #region IDisposable members
-            public void Dispose()
+            void IDisposable.Dispose()
             {
-                foreach (ISymbolReader reader in this._symbolReadersByType.Values)
+                foreach (var assemblyDefinition in this._assembliesByLocation.Values)
                 {
-                    IDisposable disposable = reader as IDisposable;
-                    if (disposable != null)
-                    {
-                        disposable.Dispose();
-                    }
+                    assemblyDefinition?.Dispose();
                 }
-                this._symbolReadersByType.Clear();
+                this._assembliesByLocation.Clear();
             }
             #endregion // IDisposable member
 
@@ -170,16 +142,16 @@
 
                 // Note: we allow checking for members declared anywhere in the hierarchy
                 // and so must open a PDB for the assembly containing the declaring type.
-                ISymbolReader reader = this[declaringType];
+                TypeDefinition typeDefinition = this[declaringType];
 
                 // Failure to find PDB short-circuits all lookups
-                if (reader != null)
+                if (typeDefinition != null)
                 {
                     // PDB knows only about methods, so this is the only one we really care about
                     MethodBase methodBase = memberInfo as MethodBase;
                     if (methodBase != null)
                     {
-                        fileName = PdbSourceFileProvider.GetFileForMethod(reader, methodBase);
+                        fileName = CecilSourceFileProvider.GetFileForMethod(typeDefinition, methodBase);
                     }
                     else
                     {
@@ -200,5 +172,45 @@
             }
         }
         #endregion // Nested classes
+    }
+
+    /// <summary>
+    /// A symbol reader which supports looking for pdb files based on symbol search path on lookup failure
+    /// </summary>
+    internal sealed class SymbolReaderProvider : ISymbolReaderProvider
+    {
+        readonly PdbReaderProvider _pdbReaderProvider = new PdbReaderProvider();
+        readonly string[] _symbolSearchPaths;
+
+        public SymbolReaderProvider(string[] symbolSearchPaths)
+        {
+            // List will contain duplicates
+            _symbolSearchPaths = symbolSearchPaths?.Distinct().ToArray();
+        }
+
+        public Mono.Cecil.Cil.ISymbolReader GetSymbolReader(ModuleDefinition module, string fileName)
+        {
+            try
+            {
+                return _pdbReaderProvider.GetSymbolReader(module, fileName);
+            }
+            catch (FileNotFoundException) when (_symbolSearchPaths != null)
+            {
+                string fileNameOnly = Path.GetFileNameWithoutExtension(fileName) + ".pdb";
+                foreach (var path in _symbolSearchPaths)
+                {
+                    var possibleName = Path.Combine(path, fileNameOnly);
+                    if (File.Exists(possibleName))
+                        return _pdbReaderProvider.GetSymbolReader(module, possibleName);
+                }
+
+                throw;
+            }
+        }
+
+        public Mono.Cecil.Cil.ISymbolReader GetSymbolReader(ModuleDefinition module, Stream symbolStream)
+        {
+            return _pdbReaderProvider.GetSymbolReader(module, symbolStream);
+        }
     }
 }
